requirements.txt
python-binance
websocket-client
requests